getTransaction is in v0
